Crate bitwise

Bitwise manipulation algorithms for Words and sequences of Words.

The algorithms have long and boring readable names that explicitly state what the algorithm does, but their documentation also contain other popular names of each algorithm as well as the hardware intrinsics they map to (if any) to make them easier to find.

The following architectures are supported using feature flags:

SSE 4.2: sse42,

BMI 1.0: bmi1,

BMI 2.0: bmi2,

ABM: abm,

TBM: tbm.

The crate llvmint is used to implement the intrinsics whenever possible. Inline assembly prevents compiler optimizations and is deliberately not used. Ideally, these intrinsics should be somehow exposed by rustc (or a similar crate to this should be part of the standard library).

Traits

Word
Bitwise manipulation algorithms for Words.

Words
Bitwise manimpulation algorithms for sequences of Words.